metrics: extract influxdb v1 logic
